requirements.txt
ollama
langchain-community
requests
python-dotenv
